+++ ChatApp/etc/Todo
@@ -48,3 +48,8 @@ Fix unread badge
 - [ ] **Google Search Console** – Integrate with Google Search Console.  
 - [ ] **Buy Domain** – Purchase and configure a custom domain.  
 - [ ] **Add Images** – Ensure all necessary images are included in production.  
+
+
+move profile photo to bottom message
+dock your messages to the side
+make it so when you hover it shows the timestamp and read of that message your hovering on
